backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.4
Flask-Migrate==4.0.4
python-dotenv==1.0.0
mysql-connector-python==8.1.0
Flask-Cors==3.0.10
openpyxl==3.1.2
PyJWT==2.8.0
prophet==1.1.5
joblib==1.4.2
matplotlib==3.9.2
openmeteo-requests==1.2.0
requests-cache==1.2.0
retry-requests==2.0.0
numpy==1.26.4
pandas==2.2.2
APScheduler==3.10.4